add support for checking data.lmdb on-disk file size to WasmTestBuilder, extract utils from benches
