+++ execution_engine_testing/test_support/Cargo.toml
@@ -14,12 +14,14 @@ license = "Apache-2.0"
 casper-execution-engine = { version = "1.4.3", path = "../../execution_engine", features = ["test-support"] }
 casper-hashing = { version = "1.4.2", path = "../../hashing" }
 casper-types = { version = "1.4.5", path = "../../types" }
+filesize = "0.2.0"
 lmdb = "0.8.0"
 log = "0.4.14"
 num-rational = "0.4.0"
 num-traits = "0.2.14"
 once_cell = "1.8.0"
 rand = "0.8.4"
+tempfile = "3"
 
 [dev-dependencies]
 version-sync = "0.9.3"
